[MAINT] remove dead code
